requirements.txt
php

pandas
boto3
pyyaml
psycopg2-binary
s3fs
